+++ HISTORY.rst
@@ -17,6 +17,7 @@
   authentication.
 - Support blank app URL namespaces. These are used to register the
   ``urlpatterns`` of encapsulated libraries as top level named URLs.
+- Add a stacked Font Awesome icon class.
 
 4.0.16 (2021-09-29)
 ===================